finish lecture 7
