+++ CS193p_Memorize/CS193p_Memorize/CardView.swift
@@ -18,8 +18,6 @@ struct CardView: View {
     }
     
     private struct Constants {
-        static let cornerRadius: CGFloat = 20
-        static let lineWeight: CGFloat = 2
         static let inset: CGFloat = 5
         struct FontSize {
             static let largest: CGFloat = 60
@@ -34,27 +32,18 @@ struct CardView: View {
     
     // main/body view
     var body: some View {
-        ZStack  {
-            let base = RoundedRectangle(cornerRadius: Constants.cornerRadius)
-            Group {
-                base.fill(Color.gray)
-                Pie(endAngle: .degrees(240))
-                    .opacity(Constants.Pie.opacity)
-                    .overlay(
-                        Text(card.content)
-                            .font(.system(size: Constants.FontSize.largest))
-                            .minimumScaleFactor(Constants.FontSize.scaleFactor)
-                            .multilineTextAlignment(.center)
-                            .padding(Constants.Pie.inset)
-                        )
-                    .padding(Constants.inset)
-            }
-            .opacity(card.isFaceUp ? 1 : 0)
-            base.fill(themeColor)
-                .stroke(Color.white, lineWidth: Constants.lineWeight)
-                .opacity(card.isFaceUp ? 0 : 1)
-        }
-        .opacity(card.isFaceUp || !card.isMatched ? 1 : 0)
+        Pie(endAngle: .degrees(240))
+            .opacity(Constants.Pie.opacity)
+            .overlay(
+                Text(card.content)
+                    .font(.system(size: Constants.FontSize.largest))
+                    .minimumScaleFactor(Constants.FontSize.scaleFactor)
+                    .multilineTextAlignment(.center)
+                    .padding(Constants.Pie.inset)
+                )
+            .padding(Constants.inset)
+            .modifier(Cardify(isFaceUp: card.isFaceUp, themeColor: themeColor))
+            .opacity(card.isFaceUp || !card.isMatched ? 1 : 0)
     }
 }
