+++ CS193p_Memorize/CS193p_Memorize/MemorizeGame.swift
@@ -67,7 +67,7 @@ struct MemorizeGame<CardContent> where CardContent: Equatable {
     }
     
     struct Card: Equatable, Identifiable, CustomDebugStringConvertible {
-        var isFaceUp = true
+        var isFaceUp = false
         var isMatched = false
         let content: CardContent
